% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diagram.R
\name{compute_diagram}
\alias{compute_diagram}
\title{Draw exclusion diagram}
\usage{
compute_diagram(list_matrices, Relation, nom)
}
\arguments{
\item{list_matrices}{list of tibble of the different exclusion matrices}

\item{Relation}{relation between the different exclusion matrices}

\item{nom}{full path and name to save the resulting diagram}
}
\value{
no return
}
\description{
Draw a diagram with from exclusion matrices
}
\examples{
list_matrices2 = NULL
list_matrices2$matrix1 = tibble::tibble(N = c(15000,12000),Label = c("TOTS",">50 anys"))
list_matrices2$matrix2 = tibble::tibble(N = c(11000,10500,8500),Label = c("A","B","C"))
list_matrices2$matrix3 = tibble::tibble(N = c(10300,7000,4200,850),Label = c("ABCD","EFGH","IJKLMN","OPQRST"))
relation = tibble::tibble(mother = c("matrix1","matrix1"), daughter = c("matrix2","matrix3"))
compute_diagram(list_matrices2,relation,"C:/Users/martics/OneDrive - Nexus365/MCToolsTest/MCToolsTest/prova67.svg")

}
