% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSurvivalMatrix.R
\name{ComputeSurvivalMatrix}
\alias{ComputeSurvivalMatrix}
\title{Compute survival function}
\usage{
ComputeSurvivalMatrix(
  data,
  StartEvent,
  OutcomeEvent,
  CensoringEvent,
  MaximumTime,
  by_pairs = FALSE,
  PairsIdentifier = "pairs_id",
  ReducedForm = FALSE,
  IndividualIdentifier = "person_id"
)
}
\arguments{
\item{data}{tibble that contains all the information}

\item{StartEvent}{names of the different variables for start event}

\item{OutcomeEvent}{names of the different variables for outcome event}

\item{CensoringEvent}{names of the different variables for censor event}

\item{MaximumTime}{maximum time for censor event}

\item{by_pairs}{censoring by pairs, default FALSE}

\item{PairsIdentifier}{identifier for pairs, default pairs_id}

\item{ReducedForm}{if false the new variables are added to data, if true only
IndividualIdentifier is conserved from initial data frame}

\item{IndividualIdentifier}{identifier for each row}
}
\value{
survival matrix
}
\description{
compute survival function indicating the variable to start
follow-up, to end follow-up and the event of interest. We have options to
include a maximum time or censor by pairs.
}
\examples{
N = 100
posibles_dates = as.Date(18628:18648,origin="1970-01-01")
dia_entrada = sample(posibles_dates, size = N, replace = TRUE)
posibles_dates = as.Date(18658:18700,origin="1970-01-01")
dia_ev1 = sample (posibles_dates, size = N, replace = TRUE)
dia_ev1[sample(1:N,75)] = NA
dia_ev2 = sample (posibles_dates, size = N, replace = TRUE)
dia_ev2[sample(1:N,75)] = NA
dia_ev3 = sample (posibles_dates, size = N, replace = TRUE)
dia_ev3[sample(1:N,75)] = NA
dia_censor = sample (posibles_dates, size = N, replace = TRUE)
ID = 1:N
data_test = tibble::tibble(ID,dia_entrada,dia_censor,dia_ev1,dia_ev2,dia_ev3)
ComputeSurvivalMatrix(data_test,"dia_entrada",c("dia_ev1","dia_ev2"),"dia_censor",ReducedForm = TRUE, IndividualIdentifier = "ID")

}
